refactor(bazinga64): Organize classes in separate files
