+++ packages/bazinga64/package.json
@@ -32,7 +32,7 @@
     "string"
   ],
   "license": "GPL-3.0",
-  "main": "dist/bazinga64.js",
+  "main": "dist/index.js",
   "name": "bazinga64",
   "repository": "https://github.com/wireapp/wire-web-packages/tree/master/packages/bazinga64",
   "scripts": {
