+++ packages/bazinga64/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "src/main"
+    "rootDir": "src"
   },
   "exclude": ["dist", "node_modules"],
   "extends": "../../tsconfig.json"
